Implements toJson function
